+++ app/src/main/java/kr/co/nottodo/presentation/mypage/view/MyPageInformationFragment.kt
@@ -14,24 +14,18 @@ import androidx.navigation.fragment.findNavController
 import kr.co.nottodo.R
 import kr.co.nottodo.data.local.SharedPreferences
 import kr.co.nottodo.databinding.FragmentMyPageInformationBinding
-import kr.co.nottodo.listeners.OnDialogDismissListener
 import kr.co.nottodo.presentation.base.fragment.DataBindingFragment
 import kr.co.nottodo.presentation.login.view.LoginActivity
 import kr.co.nottodo.presentation.login.view.LoginActivity.Companion.DID_USER_CHOOSE_TO_BE_NOTIFIED
-import kr.co.nottodo.presentation.mypage.viewmodel.MyPageInformationViewModel
+import kr.co.nottodo.presentation.mypage.viewmodel.MyPageInformationNewViewModel
 import kr.co.nottodo.util.NotTodoAmplitude.trackEvent
-import kr.co.nottodo.util.PublicString.NO_INTERNET_CONDITION_ERROR
-import kr.co.nottodo.util.showNotTodoSnackBar
-import kr.co.nottodo.util.showToast
 
 
 class MyPageInformationFragment :
-    DataBindingFragment<FragmentMyPageInformationBinding>(R.layout.fragment_my_page_information),
-    OnDialogDismissListener {
+    DataBindingFragment<FragmentMyPageInformationBinding>(R.layout.fragment_my_page_information) {
 
-    private val viewModel: MyPageInformationViewModel by viewModels()
+    private val viewModel: MyPageInformationNewViewModel by viewModels()
     private val withdrawalDialogFragment by lazy { WithdrawalDialogFragment.newInstance() }
-    private val withdrawalFeedbackDialogFragment by lazy { WithdrawalFeedbackDialogFragment.newInstance() }
     private val myPageLogoutDialogFragment by lazy { MyPageLogoutDialogFragment.newInstance() }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -40,7 +34,6 @@ class MyPageInformationFragment :
         trackEnterMyPageInformation()
         setViews()
         setClickEvents()
-        setObservers()
     }
 
     private fun trackEnterMyPageInformation() {
@@ -164,45 +157,6 @@ class MyPageInformationFragment :
         binding.ivMyPageInformationBackArrow.setOnClickListener {
             findNavController().popBackStack()
         }
-    }
-
-    private fun setObservers() {
-        setWithdrawalObserver()
-    }
-
-    private fun setWithdrawalObserver() {
-        setWithdrawalSuccessObserver()
-        setWithdrawalErrorObserver()
-    }
-
-    private fun setWithdrawalErrorObserver() {
-        viewModel.withdrawalErrorResponse.observe(viewLifecycleOwner) { errorMessage ->
-            if (errorMessage == NO_INTERNET_CONDITION_ERROR) requireContext().showNotTodoSnackBar(
-                binding.root, NO_INTERNET_CONDITION_ERROR
-            )
-            else requireContext().showToast(errorMessage)
-        }
-    }
-
-    private fun setWithdrawalSuccessObserver() {
-        viewModel.withdrawalSuccessResponse.observe(viewLifecycleOwner) {
-            withdrawalDialogFragment.dismiss()
-            withdrawalFeedbackDialogFragment.show(
-                requireActivity().supportFragmentManager, withdrawalFeedbackDialogFragment.tag
-            )
-            trackEvent(getString(R.string.complete_withdrawal))
-        }
-    }
-
-
-    override fun onDialogDismiss() {
-        logout()
-    }
-
-    private fun logout() {
-        SharedPreferences.clearForLogout()
-        startActivity(Intent(requireContext(), LoginActivity::class.java))
-        if (!requireActivity().isFinishing) requireActivity().finish()
     }
 
     override fun bindViewModelWithBinding() {
